+++ packages/fscomponents/src/lib/RNContentLoader.web.ts
@@ -1,0 +1,4 @@
+export * from 'react-content-loader';
+export { default as default } from 'react-content-loader';
+// the react-content-loader/native module just re-exports these from rn-svg
+export { Rect, Circle, Path } from 'react-native-svg';