+++ packages/fsengagement/src/lib/RNContentLoader.ts
@@ -1,0 +1,3 @@
+// tslint:disable: no-submodule-imports
+export * from 'react-content-loader/native';
+export { default as default } from 'react-content-loader/native';