+++ packages/fsproductindex/src/components/ProductIndex.tsx
@@ -1,11 +1,19 @@
 import React, { Component } from 'react';
 import { StyleProp, TextStyle, View, ViewStyle } from 'react-native';
-import { isEqual } from 'lodash-es';
+import { isEqual, omit } from 'lodash-es';
 import {
   CommerceDataSource,
   CommerceTypes
 } from '@brandingbrand/fscommerce';
-import { Loading } from '@brandingbrand/fscomponents';
+import {
+  FilterListDrilldownProps,
+  FilterListProps,
+  GridProps,
+  Loading,
+  ProductItemProps,
+  RefineActionBarProps,
+  SelectableListProps
+} from '@brandingbrand/fscomponents';
 
 import { style as S } from '../styles/ProductIndex';
 import ProductIndexGrid from './ProductIndexGrid';
@@ -17,6 +25,7 @@ import {
 
 export interface UnwrappedProductIndexProps {
   columns?: number;
+  modalCancelButton?: boolean;
   fetchProducts?: (
     productQuery?: CommerceTypes.ProductQuery
   ) => Promise<CommerceTypes.ProductIndex>;
@@ -24,63 +33,71 @@ export interface UnwrappedProductIndexProps {
   errorText?: string;
   errorTextStyle?: StyleProp<TextStyle>;
   filterHeaderTitle?: string;
-  filterListProps?: any;
-  gridProps?: any;
-  handleFilterApply?: (data: any) => void;
+  filterListProps?: Partial<FilterListProps>;
+  showDrilldownClose?: boolean;
+  gridProps?: Partial<GridProps<any>>;
+  handleFilterApply?: (data: any, info?: { isButtonPress: boolean }) => void;
   handleFilterReset?: () => void;
   onLoadComplete?: (
-    loadMore: Function,
+    loadMore: () => void,
     hasAnotherPage?: boolean,
     count?: number,
     responseCount?: number
   ) => void;
-  handleSortChange?: (data: any) => void;
+  handleSortChange?: (data: CommerceTypes.SortingOption['id']) => void;
   hideActionBar?: boolean;
-  keywords?: any;
-  listStyle?: any;
-  loadingStyle?: any;
+  listStyle?: StyleProp<ViewStyle>;
+  loadingStyle?: StyleProp<ViewStyle>;
   loadMoreButtonStyle?: StyleProp<ViewStyle>;
   loadMoreButtonTextStyle?: StyleProp<TextStyle>;
   loadMoreLoadingStyle?: StyleProp<ViewStyle>;
   modalCancelStyle?: StyleProp<ViewStyle>;
   modalHeaderStyle?: StyleProp<ViewStyle>;
   modalHeaderTextStyle?: StyleProp<TextStyle>;
-  onNavigate?: (data: any) => void;
-  productItemProps?: any;
+  onNavigate?: (data: CommerceTypes.Product) => void;
+  productItemProps?: Partial<ProductItemProps>;
   productQuery: CommerceTypes.ProductQuery;
-  refineActionBarProps?: any;
+  refineActionBarProps?: Partial<RefineActionBarProps>;
+  renderGhost?: () => JSX.Element;
+  setTitleTotalItem?: (count: number) => void;
   renderFilter?: (
-    handleFilterApply: Function,
-    handleFilterReset: Function,
+    handleFilterApply: (data: any, info?: { isButtonPress: boolean }) => void,
+    handleFilterReset: () => void,
     commerceData: CommerceTypes.ProductIndex
   ) => JSX.Element;
-  renderLoadPrev?: (loadPrev: Function, hasAnotherPage: boolean) => JSX.Element;
-  renderLoadMore?: (loadMore: Function, hasAnotherPage: boolean) => JSX.Element;
+  renderLoadPrev?: (loadPrev: () => void, hasAnotherPage: boolean) => JSX.Element;
+  renderLoadMore?: (loadMore: () => void, hasAnotherPage: boolean) => JSX.Element;
   renderLoading?: () => JSX.Element;
   renderNoResult?: (
     commerceData: CommerceTypes.ProductIndex,
-    handleFilterReset: Function
+    handleFilterReset: () => void
   ) => JSX.Element;
   renderProductItem?: (data: CommerceTypes.Product) => JSX.Element;
   renderRefineActionBar?: (
-    showFilterModal: Function,
-    showSortModal: Function,
-    commerceData: CommerceTypes.ProductIndex
+    showFilterModal: () => void,
+    showSortModal: () => void,
+    commerceData: CommerceTypes.ProductIndex,
+    applyFilters?: (selectedItems: any, info?: { isButtonPress: boolean }) => void,
+    resetFilters?: () => void
   ) => JSX.Element;
   renderSort?: (
-    handleSortChange: Function,
-    commerceData: CommerceTypes.ProductIndex
+    handleSortChange: (sortItem: CommerceTypes.SortingOption) => void,
+    commerceData: CommerceTypes.ProductIndex,
+    closeSortModal?: () => void
   ) => JSX.Element;
   sortHeaderStyle?: string;
-  sortListProps?: any;
-  FilterListDrilldownProps?: any;
-  style?: any;
+  sortListProps?: Partial<SelectableListProps>;
+  FilterListDrilldownProps?: Partial<FilterListDrilldownProps>;
+  style?: StyleProp<ViewStyle>;
   modalAnimationType?: 'none' | 'slide' | 'fade';
   modalType?: 'full-screen' | 'half-screen';
   filterType?: 'accordion' | 'drilldown';
+  refinementBlacklist?: string[];
+  selectedRefinementBlacklist?: string[];
   mergeSortToFilter?: boolean;
   filterInBackground?: boolean;
   renderModalLoading?: () => JSX.Element;
+  defaultSortOption?: string;
 }
 
 export type ProductIndexProps = UnwrappedProductIndexProps & WithProductIndexProviderProps;
@@ -101,29 +118,72 @@ export class ProductIndex extends Component<UnwrappedProductIndexProps & WithPro
   }
 
   render(): JSX.Element {
-    const { format, loadingStyle, style, commerceData } = this.props;
+    const { format, loadingStyle, style, commerceData, modalCancelButton } = this.props;
     const productIndexProps = {
       ...this.props,
       onPress: this.onPress
     };
 
     if (!commerceData) {
+      if (this.props.renderGhost) {
+        return this.props.renderGhost();
+      }
       return <Loading style={[S.loading, loadingStyle]} />;
     }
     let content = null;
 
+    // Allow setting separate lists for selected refinements
+    // or just use the same list for selected and unselected
+    const selectedRefinementBlacklist: string[] | undefined =
+      this.props.selectedRefinementBlacklist || this.props.refinementBlacklist;
+
+    const filteredCommerceData = this.getCommerceData(commerceData, selectedRefinementBlacklist);
+
     switch (format) {
-      case 'grid':
-        content = <ProductIndexGrid {...productIndexProps} />;
-        break;
       case 'list':
-        content = <ProductIndexGrid columns={1} {...productIndexProps} />;
+        content = (
+          <ProductIndexGrid
+            columns={1}
+            {...productIndexProps}
+            commerceData={filteredCommerceData}
+            headerWithCancelButton={modalCancelButton}
+          />
+        );
         break;
+      case 'grid':
       default:
-        content = <ProductIndexGrid {...productIndexProps} />;
+        content = (
+          <ProductIndexGrid
+            {...productIndexProps}
+            commerceData={filteredCommerceData}
+            headerWithCancelButton={modalCancelButton}
+          />
+        );
     }
 
     return <View style={[S.container, style]}>{content}</View>;
+  }
+
+  private getCommerceData = (
+    commerceData: CommerceTypes.ProductIndex,
+    selectedRefinementBlacklist?: string[]
+  ) => {
+    const filteredCommerceData = selectedRefinementBlacklist &&
+    commerceData.selectedRefinements ? {
+      ...commerceData,
+      selectedRefinements: commerceData.selectedRefinements &&
+        omit(commerceData.selectedRefinements, selectedRefinementBlacklist),
+      refinements: commerceData.refinements && (this.props.refinementBlacklist ?
+        commerceData.refinements.filter(refinement => {
+          return this.props.refinementBlacklist?.indexOf(refinement.id) === -1;
+        }) : commerceData.refinements)
+    } : commerceData;
+
+    if (!!filteredCommerceData.total && !!this.props.setTitleTotalItem) {
+      this.props.setTitleTotalItem(filteredCommerceData.total);
+    }
+
+    return filteredCommerceData;
   }
 }