+++ packages/pirateship/src/components/PSCreditCardForm.tsx
@@ -3,8 +3,8 @@ import { ImageRequireSource, InteractionManager, StyleProp, View, ViewStyle } fr
 import formFieldStyles from '../styles/FormField';
 import { select, textbox, textboxWithRightIcon } from '../lib/formTemplate';
 import { Form } from '@brandingbrand/fscomponents';
-// @ts-ignore TODO: Add types for tcomb-form-native
-import * as t from 'tcomb-form-native';
+// Using import with tcomb-form-native seems to cause issues with the object being undefined.
+const t = require('@brandingbrand/tcomb-form-native');
 import CCType, { CardBrand } from 'credit-card-type';
 
 const months = require('../../assets/months.json');
@@ -16,7 +16,7 @@ const icons: { [key in CardBrand]?: ImageRequireSource } = {
 };
 
 export interface FormOptions {
-  stylesheet: object;
+  stylesheet: any;
   fields: any;
 }
 
@@ -34,11 +34,11 @@ export default class PSCreditCardForm extends Component<
   PSCreditCardFormProps,
   PSCreditCardFormState
 > {
-  formRef: Form | null = null;
+  formRef: any;
 
-  formType: object;
+  formType: any;
 
-  fieldOptions: object = {
+  fieldOptions: any = {
     ccNumber: {
       factory: t.form.Textbox,
       template: textboxWithRightIcon,
@@ -109,6 +109,11 @@ export default class PSCreditCardForm extends Component<
     this.updateCCIcon(this.state.value && this.state.value.ccNumber);
   }
 
+  componentDidMount(): void {
+    // tslint:disable-next-line:ter-max-len
+    console.warn('PSCreditCardForm is deprecated and will be removed in the next version of Flagship.');
+  }
+
   render(): JSX.Element {
     return (
       <View style={this.props.style}>
@@ -126,7 +131,7 @@ export default class PSCreditCardForm extends Component<
   }
 
   focusField = (fieldName: string) => {
-    const field = this.formRef && this.formRef.getComponent(fieldName);
+    const field = this.formRef.getComponent(fieldName);
     if (!field) {
       return console.warn(`field ${fieldName} doesn't exist`);
     }
@@ -142,8 +147,8 @@ export default class PSCreditCardForm extends Component<
   }
 
   getFormRef = (ref: any) => (this.formRef = ref);
-  getValue = () => this.formRef && this.formRef.getValue();
-  validate = () => this.formRef && this.formRef.validate();
+  getValue = () => this.formRef.getValue();
+  validate = () => this.formRef.validate();
   handleChange = (value: any) => {
     const oldCC = this.state.value && this.state.value.ccNumber;
     const newCC = value.ccNumber;